[utils] `log_subprocess` → `logged`
Make `log_subprocess` decorator more generic:
* Rename it `logged`
* Move it to `utils/debug`
